fix: address edge case in autoInject comments
Unforunately, cdfb4917e6028c8f966276d6e792018c7fd2ae3c introduced issues
with certain edge cases for comments. Fixing the edge case in the
regular expression is likely to reintroduce the issue that commit was
addressing. This change replaces the regular expression entirely with
a function that uses `String.prototype` methods instead of regular
expressions. While more wordy, it is likely easier to read/understand
(and therefore easier to maintain) and it seems to be performant.